+++ src/git_secret_protector/core/settings.py
@@ -28,10 +28,9 @@ class Settings:
     magic_header: str = 'ENCRYPTED'
     storage_type: StorageType = StorageType.AWS_SSM
     config: configparser.ConfigParser = field(init=False)
-    use_gcp_default_credentials_for_project_id: bool = True
 
     def __post_init__(self):
-        self.base_dir = self.find_base_dir()
+        self.base_dir = self._find_base_dir()
         self.module_dir = os.path.join(self.base_dir, self.module_folder)
         self.config_file = os.path.join(self.module_dir, 'config.ini')
         self.cache_dir = os.path.join(self.module_dir, 'cache')
@@ -40,7 +39,8 @@ class Settings:
         self.config = configparser.ConfigParser()
         self._load_config()
 
-    def find_base_dir(self):
+    @staticmethod
+    def _find_base_dir():
         current_dir = os.getcwd()
         while current_dir != os.path.dirname(current_dir):  # Traverse up to the root directory
             possible_dir = os.path.join(current_dir, Settings.BASE_DIR_LOOKUP_FOLDER)
@@ -59,7 +59,6 @@ class Settings:
             self.log_max_size = self.config.getint('DEFAULT', 'log_max_size', fallback=self.log_max_size)
             self.log_backup_count = self.config.getint('DEFAULT', 'log_backup_count', fallback=self.log_backup_count)
             self.magic_header = self.config.get('DEFAULT', 'magic_header', fallback=self.magic_header)
-            self.use_gcp_default_credentials_for_project_id = self.config.getboolean('DEFAULT', 'use_gcp_default_credentials_for_project_id', fallback=self.use_gcp_default_credentials_for_project_id)
 
             storage_type_str = self.config.get('DEFAULT', 'storage_type', fallback=self.storage_type.value)
             if storage_type_str in [member.value for member in StorageType]: