+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 **Note**: Ensure to keep this changelog updated with every new release or change made to the project.
 
+## [Unreleased]
+
+### Changed
+
+- Removed redundant code for project ID retrieval using gcloud CLI.
+
 ## [1.1.0] - 2024-11-11
 
 ### Changed